requirements.txt
pandas
numpy
nltk
sentence-transformers
faiss-cpu
